+++ Simulator/Runtime/Source/Solvers/MCT/MCTSolver.cpp
@@ -4,6 +4,7 @@
 #include <random>
 #include <algorithm>
 #include <omp.h>
+#include <cmath>
 
 #include "Logger/Logger.hpp"
 #include "Model/ExportUtil.hpp"
@@ -49,7 +50,7 @@ namespace VT_Physics::mct {
         m_host_data->inv_dt2 = m_host_data->inv_dt * m_host_data->inv_dt;
         m_host_data->cur_simTime = 0.f;
         m_host_data->particle_radius = mct_config["Required"]["particleRadius"].get<float>();
-        m_host_data->fPart_rest_volume = std::powf(2 * m_host_data->particle_radius, 3);
+        m_host_data->fPart_rest_volume = std::pow(2 * m_host_data->particle_radius, 3);
         m_host_data->h = m_host_data->particle_radius * 4;
         m_host_data->div_free_threshold = mct_config["Required"]["divFreeThreshold"].get<float>();
         m_host_data->incomp_threshold = mct_config["Required"]["incompThreshold"].get<float>();
