+++ Simulator/Runtime/Source/Solvers/PBF/PBFSolver.cpp
@@ -1,6 +1,7 @@
 #include "Solvers/PBF/PBFSolver.hpp"
 
 #include <string>
+#include <cmath>
 
 #include "Logger/Logger.hpp"
 #include "Model/ExportUtil.hpp"
@@ -47,7 +48,7 @@ namespace VT_Physics::pbf {
         m_host_data->particle_radius = pbf_config["Required"]["particleRadius"].get<float>();
         m_host_data->fPart_rest_density = pbf_config["Required"]["fPartRestDensity"].get<float>();
         m_host_data->bPart_rest_density = pbf_config["Required"]["bPartRestDensity"].get<float>();
-        m_host_data->fPart_rest_volume = std::powf(2 * m_host_data->particle_radius, 3);
+        m_host_data->fPart_rest_volume = std::pow(2 * m_host_data->particle_radius, 3);
         m_host_data->h = m_host_data->particle_radius * 4;
         m_host_data->XSPH_k = pbf_config["Required"]["XSPH_k"].get<float>();
         m_host_data->gravity = make_float3(0.f, -9.8f, 0.f);
